requirements.txt



Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1

gunicorn==20.1.0


pandas==1.2.4

pytest==6.2.3



scikit-image==0.18.1


sklearn==0.0

pandas

unicodecsv==0.14.1

WTForms==2.3.3

xgboost==1.4.2

